Moving the core facture library into a separate file.
